refactor(feature/ujn/#45): interface 분리
