+++ yeoreum/src/types/friend.ts
@@ -1,0 +1,5 @@
+export interface FriendListType {
+  userNo: number;
+  nickname: string;
+  profileImage: string;
+}
